Using useEffect with Redux.
useSelector subscribes to the redux store, thus whenever state.cart is updated, useEffect will execute and send a PUT request with the cart data to the backend.

The cart data will have been updated through the redux reducer function, therefore we are sending the formatted cart data to the backend.